feat: refactor category data handling and update imports
