+++ src/data/roadmap.ts
@@ -145,11 +145,3 @@ export const getRoadmapWithSession = async (
     isBookmarked,
   };
 };
-
-export const getCategories = unstable_cache(
-  async (): Promise<RoadmapCategory[]> => {
-    return await db.query.categories.findMany({
-      orderBy: (fields, { asc }) => asc(fields.id),
-    });
-  },
-);
